requirements.txt
fastapi==0.110.*
sqlmodel==0.0.16
uvicorn==0.29.*
python-dotenv==1.*
pymysql==1.*
bcrypt==4.*
pyjwt==2.*
